【改修】(フロント)Deleteキーによる選択中メディア画像の削除処理を追加

+++ resources/js/components/media/media_contents/objects/figure/MediaFigureMngComponent.vue
@@ -81,9 +81,11 @@
     mounted(){
       document.addEventListener('keydown', (e)=> {
         if(e.code=="Delete"){
-          const figures_reverse = this.$refs.figures.reverse();
-          figures_reverse.forEach(figure => {figure.delete(); });
-          this.reRenderAll();
+          if(this.$refs.figures){
+            const figures_reverse = this.$refs.figures.reverse();
+            figures_reverse.forEach(figure => {figure.delete(); });
+            this.reRenderAll();
+          }
         }
       })
     },
